# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API for managing translations with multi-locale support",
        "version": "1.0.0"
    },
    "paths": {
        "/api/generate-token": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Generate or return existing bearer token",
                "description": "Creates a user (admin@example.com) if not exists, and returns the bearer token. If token already exists, retrieves it from DB.",
                "operationId": "3852f3274c6de3a7b5f4dff6c018175b",
                "responses": {
                    "200": {
                        "description": "Bearer token generated or retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|xYzAbcDefGhIjKlMnOpQrStUvWxYz"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List translations with filters",
                "operationId": "getTranslations",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "web,mobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TranslationKey"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create new translation",
                "operationId": "createTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "values"
                                ],
                                "properties": {
                                    "namespace": {
                                        "type": "string",
                                        "example": "app"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "auth.login"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Login button"
                                    },
                                    "values": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "locale": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Login"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationKey"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get a translation by ID",
                "description": "Returns a single translation including values and tags",
                "operationId": "getTranslationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationKey"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update an existing translation",
                "description": "Updates namespace, key, description, tags, and values of a translation",
                "operationId": "updateTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "namespace": {
                                        "type": "string",
                                        "example": "app"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "auth.login"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Login button"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "web"
                                        }
                                    },
                                    "values": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "locale": {
                                                    "type": "string",
                                                    "example": "fr"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Connexion"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated translation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationKey"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "description": "Deletes a translation by ID",
                "operationId": "deleteTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation deleted"
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations in JSON",
                "operationId": "exportTranslations",
                "parameters": [
                    {
                        "name": "locales",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "en,fr"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "auth.login": {
                                        "en": "Login",
                                        "fr": "Connexion"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TranslationKey": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "namespace": {
                        "type": "string",
                        "example": "app"
                    },
                    "key": {
                        "type": "string",
                        "example": "auth.login"
                    },
                    "description": {
                        "type": "string",
                        "example": "Login button"
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TranslationValue"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "web"
                        }
                    }
                },
                "type": "object"
            },
            "TranslationValue": {
                "properties": {
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "value": {
                        "type": "string",
                        "example": "Login"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}